script: Print callchains and symbols if they exist
Patchwork-id: 101795
O-Subject: [PATCH RHEL6.7 BZ1172076 033/313] perf script: Print callchains and symbols if they exist
Bugzilla: 1172076
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1172076
https://bugzilla.redhat.com/show_bug.cgi?id=1172076

upstream
========
commit 80b8b496ec6edaff01f9ab74dbe8a517cd718de8
Author: David Ahern <dsahern@gmail.com>
Date: Tue Nov 19 21:07:37 2013 -0700

description
===========
The intent of perf-script is to dump the events and information in the
file. H/W, S/W and raw events all dump callchains if they are present;
might as well make that the default for tracepoints too.

v2: Only add options for sym, dso and ip if callchains are present
Signed-off-by: Rafael Aquini <aquini@redhat.com>