symbols: Fix not finding kcore in buildid cache
Patchwork-id: 101800
O-Subject: [PATCH RHEL6.7 BZ1172076 036/313] perf symbols: Fix not finding kcore in buildid cache
Bugzilla: 1172076
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1172076
https://bugzilla.redhat.com/show_bug.cgi?id=1172076

upstream
========
commit 449867e346bfd52c5df6bba5b706a795c35e78d4
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Nov 26 15:19:24 2013 +0200

description
===========
The logic was not looking in the buildid cache for kcore if the host
kernel buildid did not match the recorded kernel buildid.

This affects the non-live case i.e. the kernel has changed and we are
looking at a special copy of kcore that we placed in the buildid cache
(using "perf buildid-cache -v -k /proc/kcore") when the data was
recorded.

After this fix kernel symbols get resolved/annotated correctly.

[ Added further explanation extracted from conversation between Ingo & Adrian on lkml ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>