symbols: Constify dso->long_name
Patchwork-id: 101886
O-Subject: [PATCH RHEL6.7 BZ1172076 119/313] perf symbols: Constify dso->long_name
Bugzilla: 1172076
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1172076
https://bugzilla.redhat.com/show_bug.cgi?id=1172076

upstream
========
commit bf4414ae7b86cddca60a5b510954a37d30583a1f
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Dec 10 15:19:23 2013 -0300

description
===========
Same reason as for dso->short_name, it may point to a const string, and
in most places it is treated as const, i.e. it is just accessed for
using its contents as a key or to show it on reports.

Signed-off-by: Rafael Aquini <aquini@redhat.com>