symbols: Fix bug in usage of the basename() function
Patchwork-id: 101889
O-Subject: [PATCH RHEL6.7 BZ1172076 123/313] perf symbols: Fix bug in usage of the basename() function
Bugzilla: 1172076
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1172076
https://bugzilla.redhat.com/show_bug.cgi?id=1172076

upstream
========
commit ac5e7f84c0e050fe19146d9bf51f69890beabcef
Author: Stephane Eranian <eranian@google.com>
Date: Thu Dec 5 19:26:42 2013 +0100

description
===========
The basename() implementation varies a lot between systems.

The Linux man page says: "basename may modify the content of the path,
so it may be desirable to pass a copy when calling the function".

On some other systems, the returned address may come from an internal
buffer which can be reused in subsequent calls, thus the results should
also be copied.

The dso__set_basename() function was not doing this causing problems
on some systems with wrong library names being shown by perf report,
such as on Android systems.

This patch fixes the problem.
The patch is relative to tip.git.

In v2, we clean up the comments based on Ingo's feedback.

[ v3: Fixed up wrt allocated flag now being set in dso__set_short_name ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>