tools: lib: traceevent: Replace tabs with spaces for all non-commands statements
Patchwork-id: 101977
O-Subject: [PATCH RHEL6.7 BZ1172076 217/313] tools lib traceevent: Replace tabs with spaces for all non-commands statements
Bugzilla: 1172076
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1172076
https://bugzilla.redhat.com/show_bug.cgi?id=1172076

upstream
========
commit 198430b56d4ab49d77381ef7cd3daf418a2c00c0
Author: Jiri Olsa <jolsa@redhat.com>
Date: Thu Jan 2 10:53:04 2014 +0100

description
===========
The tabbed indentation in non-commands statements could be sometimes
considered as follow up for the rule command in the Makefile.

This error is hard to find, so as a precaution replacing tabs with
spaces for all non-commands statements.

Link: http://marc.info/?t=136484403900003&r=1&w=2
Signed-off-by: Rafael Aquini <aquini@redhat.com>