callchain: Spare double comparison of callchain first entry
Patchwork-id: 102052
O-Subject: [PATCH RHEL6.7 BZ1172076 275/313] perf callchain: Spare double comparison of callchain first entry
Bugzilla: 1172076
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1172076
https://bugzilla.redhat.com/show_bug.cgi?id=1172076

upstream
========
commit b965bb41061ad8d3eafda6e7feef89279fcd3916
Author: Frederic Weisbecker <fweisbec@gmail.com>
Date: Tue Jan 14 16:37:15 2014 +0100

description
===========
When a new callchain child branch matches an existing one in the rbtree,
the comparison of its first entry is performed twice:

1) From append_chain_children() on branch lookup

2) If 1) reports a match, append_chain() then compares all entries of
the new branch against the matching node in the rbtree, and this
comparison includes the first entry of the new branch again.

Lets shortcut this by performing the whole comparison only from
append_chain() which then returns the result of the comparison between
the first entry of the new branch and the iterating node in the rbtree.
If the first entry matches, the lookup on the current level of siblings
stops and propagates to the children of the matching nodes.

This results in less comparisons performed by the CPU.

Signed-off-by: Rafael Aquini <aquini@redhat.com>