tools: Remove unnecessary callchain cursor state restore on unmatch
Patchwork-id: 102025
O-Subject: [PATCH RHEL6.7 BZ1172076 276/313] perf tools: Remove unnecessary callchain cursor state restore on unmatch
Bugzilla: 1172076
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1172076
https://bugzilla.redhat.com/show_bug.cgi?id=1172076

upstream
========
commit 2a29190c040c0b11e39197c67abf6f87e0a61f9a
Author: Frederic Weisbecker <fweisbec@gmail.com>
Date: Tue Jan 14 16:37:16 2014 +0100

description
===========
If a new callchain branch doesn't match a single entry of the node that
it is given against comparison in append_chain(), then the cursor is
expected to be at the same position as it was before the comparison
loop.

As such, there is no need to restore the cursor position on exit in case
of non matching branches.

Signed-off-by: Rafael Aquini <aquini@redhat.com>