tools: Fix traceevent plugin path definitions
Patchwork-id: 102079
O-Subject: [PATCH RHEL6.7 BZ1172076 285/313] perf tools: Fix traceevent plugin path definitions
Bugzilla: 1172076
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1172076
https://bugzilla.redhat.com/show_bug.cgi?id=1172076

upstream
========
commit b935a58dbff457c27fd63e1e1bb29db20b2ee6a8
Author: Josh Boyer <jwboyer@fedoraproject.org>
Date: Wed Jan 22 10:01:48 2014 -0500

description
===========
The plugindir_SQ definition contains $(prefix) which is not needed as
the $(libdir) definition already contains prefix in it.  This leads to
the path including an extra prefix in it, e.g. /usr/usr/lib64.

The -DPLUGIN_DIR defintion includes DESTDIR.  This is incorrect, as it
sets the plugin search path to include the value of DESTDIR.  DESTDIR is
a mechanism to install in a non-standard location such as a chroot or an
RPM build root.  In the RPM case, this leads to the search path being
incorrect after the resulting RPM is installed (or in some cases an RPM
build failure).

Remove both of these unnecessary inclusions.

Signed-off-by: Rafael Aquini <aquini@redhat.com>