symbols: Prevent the use of kcore if the kernel has moved
Patchwork-id: 102055
O-Subject: [PATCH RHEL6.7 BZ1172076 296/313] perf symbols: Prevent the use of kcore if the kernel has moved
Bugzilla: 1172076
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 1172076
https://bugzilla.redhat.com/show_bug.cgi?id=1172076

upstream
========
commit a00d28cb72d3629c6481fe21ba6c6b4f96caed49
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Wed Jan 29 16:14:41 2014 +0200

description
===========
Use of kcore is predicated upon it matching the recorded data.  If the
kernel has been relocated at boot time (i.e. since the data was
recorded) then do not use kcore.

Note that it is possible to make a copy of kcore at the time the data is
recorded using 'perf buildid-cache'.  Then the perf tools will use the
copy because it does match the data.

Signed-off-by: Rafael Aquini <aquini@redhat.com>