alps: add semi-MT support for v3 protocol
Patchwork-id: 102290
O-Subject: [PATCH RHEL6.7 01/33] Input: alps: add semi-MT support for v3 protocol
Bugzilla: 1109395 947488
RH-Acked-by: Hans de Goede <hdegoede@redhat.com>
RH-Acked-by: Aristeu S. Rozanski F. <arozansk@redhat.com>

Bugzillas:
  https://bugzilla.redhat.com/show_bug.cgi?id=947488
  https://bugzilla.redhat.com/show_bug.cgi?id=1109395
Upstream: since v3.3
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=8404727
Tested by me on a Toshiba Tecra W50-A with a v7 Alps touchpad,
and by destkop-QA on a Dell M4700

At the time of the previous backports, the input layer was missing
the multitouch bits, so this backport was not possible.
The patch differs a bit from the upstream because the driver has been
updated with commits from after this one.

commit 01ce661fc83005947dc958a5739c153843af8a73
Author: Seth Forshee <seth.forshee@canonical.com>
Date:   Mon Nov 7 19:54:13 2011 -0800

    Input: ALPS - add semi-MT support for v3 protocol

    Signed-off-by: Seth Forshee <seth.forshee@canonical.com>
    Acked-by: Chase Douglas <chase.douglas@canonical.com>
    Signed-off-by: Dmitry Torokhov <dtor@mail.ru>

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>