add efi_smbios_addr kernel parameter
Patchwork-id: 102492
O-Subject: [RHEL6.7 PATCH BZ 1179177 v2] x86, add efi_smbios_addr kernel parameter
Bugzilla: 1179177
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1179177
Date: Tue Jan  6 08:46:22 EST 2015
Build OS: Red Hat Enterprise Linux Server release 6.6 (Santiago)
Cross compiler version: package rhel6-cross-compilers is not installed
System name: intel-canoepass-05.lab.bos.redhat.com with -j24
Built on: kernel-2.6.32-521.el6
Arch built: i686 ppc64 s390x x86_64

RHEL_only.

Bull 20 socket, 12TBsystem will not boot a kdump kernel due to both
software and hardware issues.  On the software side, the kernel is failing
to find all the PCI segments and busses.  This occurs because of a failed
DMI check in acpi_mcfg_check_entry().  A closer look at the boot log shows
the kdump kernel did not find the SMBIOS or DMI tables.

When kdumping an efi enabled system, the following failure is seen in
the boot log:

DMI not present or invalid.

On efi enabled systems, the address of the dmi table is passed in via efi,
and stored in the efi struct as efi.smbios.  In RHEL6 we boot the kdump
kernel with 'noefi', so the efi struct is not populated and the kernel
looks for smbios at the legacy location of 0xf000, which on an efi enabled
system is not valid.

We are not backporting all the efi + kdump code into RHEL6 so I am
introducing a parameter that the kdump config can call to define the
location of the smbios, efi_smbios_addr.  The value is already exported
in sysfs, /sys/firmware/efi/systab, so that kexec-tools can read the value
there and pass it into the kdump kernel.

Successfully tested as part of debug in BZ 1155231.  As mentioned in the BZ
this also makes a large HP boot with fewer problems.

[v2]: Documentation update

Cc: Don Zickus <dzickus@redhat.com>
Cc: Tony Camuso <tcamuso@redhat.com>
Cc: lszubowi@redhat.com
Cc: vgoyal@redhat.com
Cc: dyoung@redhat.com
Cc: bhe@redhat.com
Signed-off-by: Rafael Aquini <aquini@redhat.com>