kernel: fix cpu target address of directed yield
Patchwork-id: 102572
O-Subject: [RHEL6.7 PATCH] [s390] kernel: fix cpu target address of directed yield
Bugzilla: 1180061
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Description
-----------
The System might become unresponsive or shows unexpected latencies.

The kernel incorrectly has two definitions of cpu_logical_map().
This function maps logical cpu numbers to physical cpu addresses.
For example, if a spinlock is being held by a different cpu the kernel will
try to donate its cpu time to the cpu which owns the lock.

When translating the logical cpu number of the owner of the spinlock to the
corresponding physical cpu number the incorrect second definition of
cpu_logical_map() was being used.  The second definition always uses a
one-to-one mapping of logical to physical cpu addresses. This mapping however
is likely wrong after a reboot and therefore the target address of the
directed yield call which should donate cpu time would be wrong.  Especially
if the target cpu is in stopped state due to a cpu hotplug operation this can
lead to significant latencies.

Remove the second definition of cpu_logical_map().

Bugzilla
--------
BZ 1180061
https://bugzilla.redhat.com/show_bug.cgi?id=1180061

Upstream status of the patch
----------------------------
This problem fix is specific to the RHEL6 release and not applicable to
the upstream kernel.

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=8477547

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Rafael Aquini <aquini@redhat.com>