kvm: Avoid pagefault in kvm_lapic_sync_to_vapic
Patchwork-id: 102834
O-Subject: [RHEL6.7/6.6.z/6.5.z PATCH] kvm: Avoid pagefault in kvm_lapic_sync_to_vapic
Bugzilla: 1116398
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>

This patch, done by me and Andrea, fixes a deadlock caused by the
interaction between KSM and KVM.

KVM might trigger a pagefault in kvm_lapic_sync_to_vapic while IRQs
are disabled.  Then the CPU that runs KSM, with a lock taken, waits
for other CPUs to process an IPI.  The CPU that runs KVM tries to take
the lock in the pagefault handler, but it gets stuck because IRQs are
disabled and and IPIs won't be processed.

Upstream fixed the problem by moving the write outside the IRQ-disabled
region (commit b463a6f744a2, KVM: Non-atomic interrupt injection,
2010-07-20).  This is not something we want to backport to RHEL6,
unfortunately.

Instead, use kvm_write_guest_cached_atomic and retry the vmentry
if it fails.  In order to retry the vmentry immediately, without
executing a single instruction, send an IPI to the current physical CPU.
This trick is also in use upstream for a different reason (upstream,
commit b463a6f744a2 also introduced cancel_injection and uses it to
cancel a vmentry); in both cases, the IPI occurs after the event has
been injected into the guest but before the first instruction executes.

Andrea pointed out however that, by disabling pagefaults, any COW on the
page will not be broken and will result in a livelock.  The COW could be
created by KSM---unlikely but possible.  Hence, the page is pre-faulted
before entering the irq-disabled region, by reading and then writing
back the first byte (the VAPIC region fits in a page).  This guarantees
progress, and in the common case is much faster than get_user_pages.

apic_sync_pv_eoi_to_guest already uses kvm_write_guest_cached_atomic,
and is not subject to livelock, so it does not need prefaulting.

Tested by installing Windows XP (which uses the VAPIC), RHEL4 and
Fedora 20, with both flexpriority=Y and flexpriority=N.

Signed-off-by: Rafael Aquini <aquini@redhat.com>