ipoib: Add a helper to restart the multicast task
Patchwork-id: 103398
O-Subject: [PATCH RHEL6.7 03/12] IB/ipoib: Add a helper to restart the multicast task
Bugzilla: 1159925 1184072
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Add a simple helper to do the right thing when restarting
the multicast thread.  Call that helper from all the places
that need to restart the thread.  Add two places that didn't
restart the thread, but should have.

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>