raid1: Rewrite the implementation of iobarrier
Patchwork-id: 103685
O-Subject: [RHEL6.7 PATCH 01/53] raid1: Rewrite the implementation of iobarrier.
Bugzilla: 1136260 1149935 1152914 1153341 1177913
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: majianpeng <majianpeng@gmail.com>

[re-introduce patch which was backed out last minute for RHEL 6.6]

There is an iobarrier in raid1 because of contention between normal IO and
resync IO.  It suspends all normal IO when resync/recovery happens.

However if normal IO is out side the resync window, there is no contention.
So this patch changes the barrier mechanism to only block IO that
could contend with the resync that is currently happening.

We partition the whole space into five parts.

Signed-off-by: Rafael Aquini <aquini@redhat.com>