raid5: avoid livelock caused by non-aligned writes
Patchwork-id: 103688
O-Subject: [RHEL6.7 PATCH 17/53] md/raid5: avoid livelock caused by non-aligned writes.
Bugzilla: 1136260 1149935 1152914 1153341 1177913
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

If a stripe in a raid6 array received a write to each data block while
the array is degraded, and if any of these writes to a missing device
are not page-aligned, then a live-lock happens.

In this case the P and Q blocks need to be read so that the part of
the missing block which is *not* being updated by the write can be
constructed.  Due to a logic error, these blocks are not loaded, so
the update cannot proceed and the stripe is 'handled' repeatedly in an
infinite loop.

This bug is unlikely as most writes are page aligned.  However as it
can lead to a livelock it is suitable for -stable.  It was introduced
in 3.16.

Cc: stable@vger.kernel.org (v3.16)
Fixed: 67f455486d2ea20b2d94d6adf5b9b783d079e321
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit a40687ff73a5b14909d6aa522f7d778b158911c5)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>