raid1: fix_read_error should act on all non-faulty devices
Patchwork-id: 103710
O-Subject: [RHEL6.7 PATCH 28/53] md/raid1: fix_read_error should act on all non-faulty devices.
Bugzilla: 1136260 1149935 1152914 1153341 1177913
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: NeilBrown <neilb@suse.de>

If a devices is being recovered it is not InSync and is not Faulty.

If a read error is experienced on that device, fix_read_error()
will be called, but it ignores non-InSync devices.  So it will
neither fix the error nor fail the device.

It is incorrect that fix_read_error() ignores non-InSync devices.
It should only ignore Faulty devices.  So fix it.

This became a bug when we allowed reading from a device that was being
recovered.  It is suitable for any subsequent -stable kernel.

Fixes: da8840a747c0dbf49506ec906757a6b87b9741e9
Cc: stable@vger.kernel.org (v3.5+)
Reported-by: Alexander Lyakas <alex.bolshoy@gmail.com>
Tested-by: Alexander Lyakas <alex.bolshoy@gmail.com>
Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit b8cb6b4c121e1bf1963c16ed69e7adcb1bc301cd)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>