tools: Fix synthesizing mmaps for threads
Patchwork-id: 104065
O-Subject: [PATCH RHEL6.7 BZ1188336 062/280] perf tools: Fix synthesizing mmaps for threads
Bugzilla: 1188336
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1188336
https://bugzilla.redhat.com/show_bug.cgi?id=1188336

upstream
========
commit bfd66cc71a3f831ba7c2116d79416cfb8883f6cf
Author: Don Zickus <dzickus@redhat.com>
Date: Wed Feb 26 10:45:26 2014 -0500

description
===========
Currently if a process creates a bunch of threads using pthread_create
and then perf is run in system_wide mode, the mmaps for those threads
are not captured with a synthesized mmap event.

The reason is those threads are not visible when walking the /proc/
directory looking for /proc/<pid>/maps files.  Instead they are
discovered using the /proc/<pid>/tasks file (which the synthesized comm
event uses).

This causes problems when a program is trying to map a data address to a
tid.  Because the tid has no maps, the event is dropped.  Changing the
program to look up using the pid instead of the tid, finds the correct
maps but creates ugly hacks in the program to carry the correct tid
around.

Fix this by moving the walking of the /proc/<pid>/tasks up a level (out
of the comm function) based on Arnaldo's suggestion.

Tweaked things a bit to special case the 'full' bit and 'guest' check.

Signed-off-by: Rafael Aquini <aquini@redhat.com>