session: Change header.misc dump from decimal to hex
Patchwork-id: 104070
O-Subject: [PATCH RHEL6.7 BZ1188336 067/280] perf session: Change header.misc dump from decimal to hex
Bugzilla: 1188336
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1188336
https://bugzilla.redhat.com/show_bug.cgi?id=1188336

upstream
========
commit 0ea590ae8198547d5898c72b04fa9d8f23bd0b8f
Author: Don Zickus <dzickus@redhat.com>
Date: Tue Feb 25 22:43:46 2014 -0500

description
===========
When printing the raw dump of a data file, the header.misc is
printed as a decimal.  Unfortunately, that field is a bit mask, so
it is hard to interpret as a decimal.

Print in hex, so the user can easily see what bits are set and more
importantly what type of info it is conveying.

V2: add 0x in front per Jiri Olsa
Signed-off-by: Rafael Aquini <aquini@redhat.com>