symbols: Record the reason for filtering an address_location
Patchwork-id: 104078
O-Subject: [PATCH RHEL6.7 BZ1188336 075/280] perf symbols: Record the reason for filtering an address_location
Bugzilla: 1188336
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1188336
https://bugzilla.redhat.com/show_bug.cgi?id=1188336

upstream
========
commit b3cef7f60f17d953545f7069f6407fc24202a64d
Author: Namhyung Kim <namhyung@kernel.org>
Date: Mon Mar 17 16:59:21 2014 -0300

description
===========
By turning the addr_location->filtered member from a boolean to a u8
bitmap, reusing (and extending) the hist_filter enum for that.

This patch doesn't change the logic at all, as it keeps the meaning of
al->filtered !0 to mean that the entry _was_ filtered, so no change in
how this value is interpreted needs to be done at this point.

This will be soon used in upcoming patches.

[ yanked this out of a previous patch ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>