symbols: Apply all filters to an addr_location
Patchwork-id: 104079
O-Subject: [PATCH RHEL6.7 BZ1188336 076/280] perf symbols: Apply all filters to an addr_location
Bugzilla: 1188336
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1188336
https://bugzilla.redhat.com/show_bug.cgi?id=1188336

upstream
========
commit 466fa7647413665dcba46c4f7f2b4a9808426989
Author: Namhyung Kim <namhyung@kernel.org>
Date: Mon Mar 17 17:12:27 2014 -0300

description
===========
Instead of bailing out as soon as we find a filter that applies, go on
checking all of them so that we can zoom in/out filters.

We also need to make sure we only update al->filtered after
thread__find_addr_map(), because there is where al->filtered gets
initialized to zero.

This will increase the cost of processing when all we don't need this
toggling, but will provide flexibility for the TUI and GTK+ interfaces,
that will incur in creating the hist_entries just once.

[ yanked this out of a previous patch ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>