tools: Allow ability to map cpus to nodes easily
Patchwork-id: 104139
O-Subject: [PATCH RHEL6.7 BZ1188336 130/280] perf tools: Allow ability to map cpus to nodes easily
Bugzilla: 1188336
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1188336
https://bugzilla.redhat.com/show_bug.cgi?id=1188336

upstream
========
commit 7780c25bae59fd042281710f5e1243268e2c18dc
Author: Don Zickus <dzickus@redhat.com>
Date: Mon Apr 7 14:55:21 2014 -0400

description
===========
This patch figures out the max number of cpus and nodes that are on the
system and creates a map of cpu to node.  This allows us to provide a cpu
and quickly get the node associated with it.

It was mostly copied from builtin-kmem.c and tweaked slightly to use less memory
(use possible cpus instead of max).  It also calculates the max number of nodes.

[ Removing out label code in init_cpunode_map ]
[ Adding check for snprintf error ]
[ Removing unneeded returns ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>