alsa/hda/usb/core: hda - Enable mute/mic-mute LEDs for more Thinkpads with Realtek codec
Patchwork-id: 104670
O-Subject: [RHEL 6.7 PATCH 203/846] ALSA/HDA/USB/core: hda - Enable mute/mic-mute LEDs for more Thinkpads with Realtek codec
Bugzilla: 1069236
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

We're using the ACPI interface to detect whether we're dealing with a Thinkpad
or not. This way we're not loading the thinkpad_acpi module when we're not on
a Thinkpad, but at the same time, we give the opportunity to check for, and
potentially enable, both present and future Thinkpad with mute/micmute LEDs.

At least those running the ALC269 family (269 to 299) of Realtek codecs.

Cc: Alex Hung <alex.hung@canonical.com>
Tested-by: Hui Wang <hui.wang@canonical.com>
Signed-off-by: David Henningsson <david.henningsson@canonical.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 2793769f4450d9e4845257c953ae278ecdff7d94)
rhbz#1069236
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>