alsa/hda/usb/core: usb-audio: Playback and MIDI support for Novation Twitch DJ controller
Patchwork-id: 105124
O-Subject: [RHEL 6.7 PATCH 657/846] ALSA/HDA/USB/core: usb-audio: Playback and MIDI support for Novation Twitch DJ controller
Bugzilla: 1069236
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The hardware also has a PCM capture device which is not implemented in
this patch.

It may be possible to generalise this to Saffire 6 USB support and some
of the other Focusrite interfaces, but as I don't have access to these
devices we should wait until capture support is working first.

Capture support is not implemented because the code assumes the endpoint
to have its own interface (instead, it shares the interface with playback)
and some thought will be needed to lift this limitation.

Signed-off-by: Mark Hills <mark@xwax.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 5e212332cc7eed0ffbf91fbe5bab6e2a44b83de6)
rhbz#1069236
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>