alsa/hda/usb/core: usb-audio: uac2: try to find and switch to valid clock
Patchwork-id: 105140
O-Subject: [RHEL 6.7 PATCH 673/846] ALSA/HDA/USB/core: usb-audio: UAC2: try to find and switch to valid clock
Bugzilla: 1069236
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

If a selector is available on a device, it may be pointing to a
clock source which is currently invalid.
If there is a valid clock source which can be selected, switch
to it.

Signed-off-by: Eldad Zack <eldad@fogrefinery.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 8c55af3f69be6021631628b968cad301d3f294c8)
rhbz#1069236
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>