alsa/hda/usb/core: usb - Don't trust the channel config if the channel count changed
Patchwork-id: 105224
O-Subject: [RHEL 6.7 PATCH 757/846] ALSA/HDA/USB/core: usb - Don't trust the channel config if the channel count changed
Bugzilla: 1069236
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

In case the channel count of the input terminal is not the same as
the channel count of the streaming descriptor, the channel config of
the input terminal can not be trusted. Instead fall back to a default
(guessed) channel map.

This was found on a Logitech USB Headset.

Signed-off-by: David Henningsson <david.henningsson@canonical.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 504333df8ba5fc310260285a22ab5d7cf3208795)
rhbz#1069236
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>