nfs: Be less aggressive about returning delegations for open files
Patchwork-id: 105946
O-Subject: [RHEL6.7 PATCH] NFSv4: Be less aggressive about returning delegations for open files
Bugzilla: 1145334
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

Currently, if the application that holds the file open isn't doing
I/O, we may end up returning the delegation. This means that we can
no longer cache the file as aggressively, and often also that we
multiply the state that both the server and the client needs to track.

This patch adds a check for open files to the routine that scans
for delegations that are unreferenced.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Upstream Id: b757144fd77cf5512f5b60179ba5ca8dcc5184b4
Bugzilla: 1145334

Signed-off-by: Rafael Aquini <aquini@redhat.com>