kernel: 3215 tty close crash
Patchwork-id: 106244
O-Subject: [RHEL6.7 PATCH] [s390] kernel: 3215 tty close crash
Bugzilla: 1186682
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Description
-----------
Closing the 3215 tty can cause a kernel oops because of a NULL pointer.

The 3215 tty shutdown function tty3215_close first sets the
tty pointer to NULL and then stops the tty wakeup tasklet.
The raw3215_wakeup function checks the tty pointer for NULL
but does not have a barrier between the check and the call
to tty_wakeup. This opens a small race window where the
tty3215_close function can clear the tty pointer and
tty_wakeup is called with a NULL pointer.

Reorder the code in tty3215_close to first kill the tasklet
and then set the tty pointer to NULL. Add a barrier to
raw3215_wakeup to stop the compiler from reading raw->tty
twice and add an additional check to test tty->closing.

Bugzilla
--------
BZ 1186682
https://bugzilla.redhat.com/show_bug.cgi?id=1186682

Upstream status of the patch
----------------------------
This problem is specific to RHEL6.  The upstream version is based on tty
ports which provide a better reference counting compared to the code base
in RHEL6.

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=8753232

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Rafael Aquini <aquini@redhat.com>