gfs2: Allocate reservation during splice_write
Patchwork-id: 106252
O-Subject: [RHEL6.7 PATCH] GFS2: Allocate reservation during splice_write
Bugzilla: 1193559
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

Hi,

This patch is for a bug that causes the kernel to crash whenever
sendfile() operations are done on GFS2. It was tested on system
gfs-a16c-03 using a reliable reproducer program in the bugzilla.
A patch was sent upstream to fix the problem, but due to changes
in the upstream kernel, the patch is slightly different. It has
not been accepted upstream yet, but it's pretty straightforward.

Patch text:
This patch adds a GFS2-specific function for splice_write which
first calls function gfs2_rs_alloc to make sure a reservation
structure has been allocated before attempting to reserve blocks.

rhbz#1193559

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>